requirements.txt
Flask
selenium
python-dotenv
webdriver-manager
gunicorn
